Buscar por id na ordem de compra no swagger foi arrumado, e o tipo do prazoEntrega no java foi mudado para ficar igual ao banco

+++ CRUD/src/main/java/sptech/school/CRUD/Controller/OrdemDeCompraController.java
@@ -46,7 +46,7 @@ public class OrdemDeCompraController {
 
     @GetMapping("/{id}")
     @SecurityRequirement(name = "Bearer")
-    public ResponseEntity<ListagemOrdemDeCompra> buscarPorId(@PathVariable Long id) {
+    public ResponseEntity<ListagemOrdemDeCompra> buscarPorId(@PathVariable Integer id) {
         OrdemDeCompraModel ordem = ordemDeCompraService.getById(id);
         return ResponseEntity.ok(OrdemDeCompraMapper.toListagemDto(ordem));
     }
