+++ CRUD/src/main/java/sptech/school/CRUD/dto/OrdemDeCompra/ListagemOrdemDeCompra.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Getter
@@ -13,7 +14,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class ListagemOrdemDeCompra {
 
-    private Long id;
+    private Integer id;
     private String prazoEntrega;
     private String ie;
     private String condPagamento;
